refactor: reorganize JS utilities to src/js folder
- Move validation.js and generate-schemas.js to src/js/
- Update all import paths and script references
- Update documentation to reflect new structure
