+++ package.json
@@ -13,7 +13,7 @@
     "lint": "eslint --fix --ext .js,.json .",
     "build": "node esbuild.js",
     "build:icons": "node esbuild.js --icons",
-    "schemas:generate": "node scripts/generate-schemas.js",
+    "schemas:generate": "node src/js/generate-schemas.js",
     "schemas:validate": "node scripts/validate-component.js",
     "test": "vitest",
     "test:ui": "vitest --ui",
