+++ src/js/generate-schemas.js
